requirements.txt
google-generativeai>=0.2.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
colorama==0.4.6
firecrawl-py